WIP: generate_tensor/_matrix
