+++ python/smurff/__init__.py
@@ -10,3 +10,4 @@ from .center import center_and_scale
 from .wrapper import version
 from .smurff import *
 from . import matrix_io
+from . import generate
